refactor: Rename functions that get peer/vpn data from the server.  Make it private.  Move the DB init into the InMemoryDatastore.
