+++ src/tests/test_vpn_interface.py
@@ -102,7 +102,7 @@ class TestVpnInterface:
 
             # Validate Results
             assert response.status_code == HTTPStatus.BAD_REQUEST
-            all_vpns = mock_dynamo_db.get_all_vpns()
+            all_vpns = mock_dynamo_db._get_all_vpn_from_server()
             assert all_vpns == []
 
     @patch("server_manager.ssh.paramiko.RSAKey", MagicMock())
@@ -138,7 +138,7 @@ class TestVpnInterface:
 
             # Validate Results
             assert response.status_code == HTTPStatus.BAD_REQUEST
-            all_vpns = mock_dynamo_db.get_all_vpns()
+            all_vpns = mock_dynamo_db._get_all_vpn_from_server()
             assert all_vpns == []
 
     def test_add_server_invalid_ip_network(
@@ -161,7 +161,7 @@ class TestVpnInterface:
 
         # Validate Results
         assert response.status_code == HTTPStatus.BAD_REQUEST
-        all_vpns = mock_dynamo_db.get_all_vpns()
+        all_vpns = mock_dynamo_db._get_all_vpn_from_server()
         assert all_vpns == []
 
     @patch("server_manager.ssh.paramiko.RSAKey", MagicMock())
@@ -193,7 +193,7 @@ class TestVpnInterface:
 
         # Validate Results
         assert response.status_code == HTTPStatus.OK
-        all_vpns = mock_dynamo_db.get_all_vpns()
+        all_vpns = mock_dynamo_db._get_all_vpn_from_server()
         assert all_vpns == [vpn]
 
     def test_add_server_existing_name(
@@ -217,7 +217,7 @@ class TestVpnInterface:
 
         # Validate Results
         assert response.status_code == HTTPStatus.CONFLICT
-        all_vpns = mock_dynamo_db.get_all_vpns()
+        all_vpns = mock_dynamo_db._get_all_vpn_from_server()
         assert all_vpns == [test_input]
 
     def test_add_server_existing_public_key(
@@ -240,7 +240,7 @@ class TestVpnInterface:
 
         # Validate Results
         assert response.status_code == HTTPStatus.CONFLICT
-        all_vpns = mock_dynamo_db.get_all_vpns()
+        all_vpns = mock_dynamo_db._get_all_vpn_from_server()
         assert all_vpns == [test_input]
 
     def test_get_vpn(self, mock_vpn_table, mock_peer_table, mock_vpn_manager, test_input):
@@ -313,7 +313,7 @@ class TestVpnInterface:
             response = client.get(f"/vpn/{vpn.name}?hide_secrets=false")
             updated_vpn = VpnModel(**response.json())
             assert updated_vpn.connection_info is None
-            all_vpns = mock_dynamo_db.get_all_vpns()
+            all_vpns = mock_dynamo_db._get_all_vpn_from_server()
             assert all_vpns == [updated_vpn]
 
     @patch("server_manager.ssh.paramiko.RSAKey", MagicMock())
@@ -350,7 +350,7 @@ class TestVpnInterface:
 
             # Validate Results
             assert response.status_code == HTTPStatus.BAD_REQUEST
-            all_vpns = mock_dynamo_db.get_all_vpns()
+            all_vpns = mock_dynamo_db._get_all_vpn_from_server()
             assert all_vpns == [expected_vpn]
 
     @patch("server_manager.ssh.paramiko.RSAKey", MagicMock())
@@ -383,7 +383,7 @@ class TestVpnInterface:
 
             # Validate Results
             assert response.status_code == HTTPStatus.BAD_REQUEST
-            all_vpns = mock_dynamo_db.get_all_vpns()
+            all_vpns = mock_dynamo_db._get_all_vpn_from_server()
             assert all_vpns == [expected_vpn]
 
             # Break down test
@@ -424,7 +424,7 @@ class TestVpnInterface:
             response = client.get(f"/vpn/{vpn.name}?hide_secrets=false")
             updated_vpn = VpnModel(**response.json())
             assert updated_vpn.connection_info == vpn.connection_info
-            all_vpns = mock_dynamo_db.get_all_vpns()
+            all_vpns = mock_dynamo_db._get_all_vpn_from_server()
             assert all_vpns == [test_input]
 
     def test_delete_vpn(self, mock_vpn_table, mock_peer_table, mock_vpn_manager, mock_dynamo_db, test_input):
@@ -447,5 +447,5 @@ class TestVpnInterface:
 
         # Validate Results
         assert response.status_code == 200
-        all_vpns = mock_dynamo_db.get_all_vpns()
+        all_vpns = mock_dynamo_db._get_all_vpn_from_server()
         assert all_vpns == []
